requirements.txt
pandas==2.2.0
pymongo==4.6.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1